wip: seperate login/signup views
add carousel on startup screen
