+++ src/router/index.ts
@@ -5,7 +5,8 @@ import {
   createWebHashHistory,
 } from "vue-router";
 import MainView from "@/views/main-view/MainView.vue";
-import WelcomeView from "@/views/welcome-view/WelcomeView.vue";
+import SignUp from "@/views/signup/SignUp.vue";
+import LogIn from "@/views/login/LogIn.vue";
 import CommunityView from "@/views/community-view/CommunityView.vue";
 import CommunityWelcomeView from "@/views/community-view/CommunityWelcomeView.vue";
 import ChannelView from "@/views/channel-view/ChannelView.vue";
@@ -14,7 +15,12 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: "/signup",
     name: "signup",
-    component: WelcomeView,
+    component: SignUp,
+  },
+  {
+    path: "/login",
+    name: "login",
+    component: LogIn,
   },
   {
     path: "/",
